Added new base class in models
Merged Basemixin class of scgen and trvae into one new dir called base.
Added functions that are needed in all models to base/_utils.py
